fix(i18n): re-land the catalog checker at its post-refactor home
The refactor relocated check-i18n.mjs to packages/app-core/scripts/
(rename detection was skipped — #17695 changed 25k files — so it
presented as deleted; the factory's own cycle forensics caught the
move via git ls-tree). This re-lands the advisory-hardened checker AT
THAT PATH with two fixes the move surfaced: the repo-root resolve
gains the extra hop, and the direct-run predicate adopts
pathToFileURL (the same defect class #17677 fixed — the old
`file://${argv[1]}` form breaks on spaced paths everywhere and on
Windows always). test.yml wiring repointed; the test suite moves
alongside and re-roots.

Catalogs are restored as PURE ADDITIONS over develop's live copies
(+116 keys per locale, en +847; the one true orphan removed from
every locale). The checker then caught 26 real gaps on live develop,
all fixed from their own sources: 24 missing keys filled from
call-site defaultValue / metadata defaultLabel+defaultTitle, and the
two state-dependent ternary call sites (cloudOverview localMode)
split into four distinct keys — one key can never carry two strings.

Checker: OK — 4434 literal keys, 8185 source keys × 8 locales.
Suite: 12/12 at the new path.
