+++ packages/app-core/scripts/i18n-dynamic-keys.json
@@ -1,5 +1,5 @@
 {
-  "_comment": "Keys/prefixes referenced via dynamic t(expr) or t(`prefix.${x}`) calls. The i18n linter (packages/app-core/scripts/check-i18n.mjs) treats these as 'used' when reporting orphaned locale keys. Add a literal key to 'keys' or a string prefix to 'prefixes' for each runtime-resolved translation.",
+  "_comment": "Keys/prefixes referenced via dynamic t(expr) or t(`prefix.${x}`) calls. The i18n linter (packages/scripts/check-i18n.mjs) treats these as 'used' when reporting orphaned locale keys. Add a literal key to 'keys' or a string prefix to 'prefixes' for each runtime-resolved translation.",
   "keys": [],
   "prefixes": [
     "accounts.",
@@ -58,7 +58,7 @@
     "inventory.",
     "inventoryview.",
     "knowledge.",
-    "documentsview.",
+    "knowledgeview.",
     "lifeopsoverview.",
     "lifeopspage.",
     "lifeopssettings.",
@@ -71,13 +71,11 @@
     "memory.",
     "memoryviewer.",
     "messagecontent.",
-    "elizaos.",
     "music.",
     "musicplayersettings.",
     "nav.",
     "nodeCatalog.",
     "notice.",
-    "setup.",
     "pairingview.",
     "permissionssection.",
     "pluginsview.",
@@ -121,5 +119,15 @@
     "whatsappqroverlay.",
     "workflowGraph.",
     "workspace."
+  ],
+  "uncatalogued": [
+    {
+      "key": "settings.cloudOverview.localModeLabel",
+      "reason": "call site passes a runtime-conditional defaultValue (cloud connected vs local); cataloging would freeze one branch — needs a per-branch key split in CloudOverviewSection.tsx (#17605)"
+    },
+    {
+      "key": "settings.cloudOverview.localModeDescription",
+      "reason": "same conditional-defaultValue site as localModeLabel (#17605)"
+    }
   ]
 }
