--- packages/scripts/i18n-dynamic-keys.json
@@ -1,133 +1,0 @@
-{
-  "_comment": "Keys/prefixes referenced via dynamic t(expr) or t(`prefix.${x}`) calls. The i18n linter (packages/scripts/check-i18n.mjs) treats these as 'used' when reporting orphaned locale keys. Add a literal key to 'keys' or a string prefix to 'prefixes' for each runtime-resolved translation.",
-  "keys": [],
-  "prefixes": [
-    "accounts.",
-    "advancedpageview.",
-    "agent.",
-    "agentactivitybox.",
-    "apikeyconfig.",
-    "appsview.",
-    "aria.",
-    "automations.",
-    "avatarselector.",
-    "feed.",
-    "browserworkspace.",
-    "bugreportmodal.",
-    "character.",
-    "charactereditor.",
-    "characterroster.",
-    "chat.",
-    "chatmessage.",
-    "chatsidebar.",
-    "chatview.",
-    "cloud.",
-    "cloudsetup.",
-    "cloudsourcecontrols.",
-    "codingagentcontrolchip.",
-    "codingagentsettingssection.",
-    "codingagenttaskspanel.",
-    "commandpalette.",
-    "common.",
-    "companion.",
-    "computeruseapprovaloverlay.",
-    "config-field.",
-    "config-renderer.",
-    "configpageview.",
-    "connectionfailedbanner.",
-    "connectionlostoverlay.",
-    "connectors.",
-    "conversations.",
-    "customactioneditor.",
-    "customactionspanel.",
-    "customactionsview.",
-    "databaseview.",
-    "desktopworkspacesection.",
-    "detachedshell.",
-    "elizaclouddashboard.",
-    "emotepicker.",
-    "finetuningview.",
-    "flaminaguide.",
-    "format.",
-    "game.",
-    "gameview.",
-    "gameviewoverlay.",
-    "header.",
-    "inboxview.",
-    "inference.",
-    "inventory.",
-    "inventoryview.",
-    "knowledge.",
-    "knowledgeview.",
-    "lifeopsoverview.",
-    "lifeopspage.",
-    "lifeopssettings.",
-    "lifeopsworkspace.",
-    "loadingscreen.",
-    "logspageview.",
-    "logsview.",
-    "mediagalleryview.",
-    "mediasettingssection.",
-    "memory.",
-    "memoryviewer.",
-    "messagecontent.",
-    "music.",
-    "musicplayersettings.",
-    "nav.",
-    "nodeCatalog.",
-    "notice.",
-    "pairingview.",
-    "permissionssection.",
-    "pluginsview.",
-    "policy.",
-    "portfolioheader.",
-    "provider.",
-    "providerswitcher.",
-    "ptyconsolebase.",
-    "relationships.",
-    "releasecenter.",
-    "releasecenterview.",
-    "restartbanner.",
-    "runtime.",
-    "runtimeview.",
-    "savecommandmodal.",
-    "secretsview.",
-    "settings.",
-    "settingsview.",
-    "shortcutsoverlay.",
-    "signalqroverlay.",
-    "skillsview.",
-    "startupfailureview.",
-    "startupshell.",
-    "statusbar.",
-    "streamsettings.",
-    "streamview.",
-    "subscriptionstatus.",
-    "taskseventspanel.",
-    "terminal.",
-    "todo.",
-    "tokenstable.",
-    "trajectoriesview.",
-    "trajectorydetailview.",
-    "triggers.",
-    "triggersview.",
-    "ui-renderer.",
-    "vectorbrowserview.",
-    "voice.",
-    "voiceconfigview.",
-    "wallet.",
-    "whatsappqroverlay.",
-    "workflowGraph.",
-    "workspace."
-  ],
-  "uncatalogued": [
-    {
-      "key": "settings.cloudOverview.localModeLabel",
-      "reason": "call site passes a runtime-conditional defaultValue (cloud connected vs local); cataloging would freeze one branch — needs a per-branch key split in CloudOverviewSection.tsx (#17605)"
-    },
-    {
-      "key": "settings.cloudOverview.localModeDescription",
-      "reason": "same conditional-defaultValue site as localModeLabel (#17605)"
-    }
-  ]
-}
